+++ index.php
@@ -29,7 +29,7 @@ $id     = required_param('id', PARAM_INT);
 $mode   = optional_param('mode', ROSTER_MODE_DISPLAY, PARAM_TEXT);
 $group  = optional_param('group', 0, PARAM_INT);
 $role   = optional_param('role', 0, PARAM_INT);
-$size   = optional_param('size', 100, PARAM_INT);
+$size   = optional_param('size', get_config('report_roster', 'size_default'), PARAM_INT);
 $course = $DB->get_record('course', array('id' => $id), '*', MUST_EXIST);
 
 require_login($course);