+++ locallib.php
@@ -74,11 +74,18 @@ function report_roster_get_options_role($id) {
  * @return array The user image size options
  */
 function report_roster_get_options_size() {
-    return array(
-        100 => 'S',
-        200 => 'M',
-        300 => 'L'
-    );
+    $sizes = array();
+
+    foreach(array('small', 'medium', 'large') as $size) {
+        $pixels = (int) get_config('report_roster', "size_$size");
+        $label  = get_string("size:$size", 'report_roster');
+
+        if ($pixels > 0) {
+            $sizes[$pixels] = $label;
+        }
+    }
+
+    return $sizes;
 }
 
 /**
@@ -102,16 +109,18 @@ function report_roster_output_action_buttons($id, $url, $params) {
 
     $selects = array();
     foreach($params as $key => $val) {
-        $myurl      = clone $url;
-        $myparams   = $params;
+        if (array_key_exists($key, $options) && !empty($options[$key])) {
+            $myurl      = clone $url;
+            $myparams   = $params;
 
-        unset($myparams[$key]);
-        $myurl->params($myparams);
+            unset($myparams[$key]);
+            $myurl->params($myparams);
 
-        $myselect        = new single_select($myurl, $key, $options[$key], $params[$key], null);
-        $myselect->label = get_string_manager()->string_exists("param:$key", 'report_roster') ? get_string("param:$key", 'report_roster') : get_string($key);
+            $myselect        = new single_select($myurl, $key, $options[$key], $params[$key], null);
+            $myselect->label = get_string_manager()->string_exists("param:$key", 'report_roster') ? get_string("param:$key", 'report_roster') : get_string($key);
 
-        $selects[$key] = $myselect;
+            $selects[$key] = $myselect;
+        }
     }
 
     $html = html_writer::start_tag('div');